requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.6.0
beautifulsoup4==4.13.3
blivet==3.11.0
blivet-gui==2.6.0
boto3==1.37.25
botocore==1.37.25
Brlapi==0.8.5
certifi==2025.1.31
cffi==1.17.0
charset-normalizer==3.4.1
click==8.1.8
crypt_r==3.13.1
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distro==1.9.0
dnf==4.23.0
dotenv==0.9.9
dotenv-python==0.0.1
evdev==1.9.1
fastapi~=0.115.12
fedora-third-party==0.10
file-magic==0.4.0
h11==0.14.0
humanize==4.9.0
idna==3.10
initial-setup==0.3.101
iso639==0.1.4
jmespath==1.0.1
langtable==0.0.68
legacy-cgi==2.6.2
libcomps==0.1.21
libdnf==0.74.0
louis==3.28.0
lutris==0.5.19
lxml==5.2.1
moddb==0.12.0
nftables==0.1
olefile==0.47
packaging==24.2
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
ply==3.11
productmd==1.43
protobuf==3.19.6
psutil==5.9.8
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycups==2.0.4
pydantic==2.10.6
pydantic_core==2.27.2
pyenchant==3.2.2
pygdbmi==0.11.0.0
PyGObject==3.48.2
pyinotify==0.9.6
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pykickstart==3.58
pyparted==3.13.0
pypresence==4.3.0
PyQt5==5.15.10
PyQt5_sip==12.15.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-linux-procfs==0.7.3
python-meh==0.52
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rpm==4.20.1
s3transfer==0.11.4
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sentry-sdk==2.17.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==76.0.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.8.2
soupsieve==2.6
starlette==0.46.1
systemd-python==235
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn~=0.34.2
xkbregistry==0.3

pyserial~=3.5